requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.18.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
nltk>=3.8.0
pyyaml>=6.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.32.0
requests>=2.31.0
httpx>=0.27.0
pytest>=8.0.0
ruff>=0.3.0
